Rename "Parser" namespace to "MT940"
Because it only contains MT940-related classes and the general FinTs wire-format parser is in the "Syntax" namespace.  Also move the MT940Exception out of its sub-namespace, as having a separate namespace just for exceptions isn't useful.
