+++ lib/Fhp/MT940/MT940Exception.php
@@ -1,0 +1,10 @@
+<?php
+
+namespace Fhp\MT940;
+
+/**
+ * Thrown for MT940-specific parsing errors.
+ */
+class MT940Exception extends \Exception
+{
+}
