--- lib/Fhp/Parser/Exception/MT940Exception.php
@@ -1,12 +1,0 @@
-<?php
-
-namespace Fhp\Parser\Exception;
-
-/**
- * Class MT940Exception
- * @package Fhp\Parser\Exception
- */
-class MT940Exception extends \Exception
-{
-
-}
